baseline/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-exporter-otlp-proto-http>=1.29.0
openinference-instrumentation-langchain
